[DAC] Regenerated demo for dac_dac_wav_gen_dma_sam_c21n_xpro
MH3-48669

+++ apps/dac/dac_wav_gen_dma/firmware/src/config/sam_c21n_xpro/peripheral/dac/plib_dac.c
@@ -48,9 +48,9 @@
 // *****************************************************************************
 /* This section lists the other files that are included in this file.
 */
-
-#include "plib_dac.h"
 #include "device.h"
+#include "interrupts.h"
+#include "plib_dac.h"
 
 
 /* (DAC DATA) Mask DATA[15:10] Bit */
@@ -60,12 +60,12 @@
 void DAC_Initialize(void)
 {
     /* Set Reference Voltage */
-    DAC_REGS->DAC_CTRLB = DAC_CTRLB_REFSEL(0) | DAC_CTRLB_EOEN_Msk ;
+    DAC_REGS->DAC_CTRLB = (uint8_t)(DAC_CTRLB_REFSEL(0UL) | DAC_CTRLB_EOEN_Msk );
 
-    DAC_REGS->DAC_EVCTRL = 1;
+    DAC_REGS->DAC_EVCTRL = 1U;
     
     /* Enable DAC */
-    DAC_REGS->DAC_CTRLA = DAC_CTRLA_ENABLE_Msk  ;	
+    DAC_REGS->DAC_CTRLA =(uint8_t)(DAC_CTRLA_ENABLE_Msk);	
     while((DAC_REGS->DAC_SYNCBUSY & DAC_SYNCBUSY_ENABLE_Msk) == DAC_SYNCBUSY_ENABLE_Msk)
     {
         /* Wait for Synchronization after Enabling DAC */
@@ -76,7 +76,7 @@ void DAC_Initialize(void)
 void DAC_DataWrite(uint16_t data)
 {
     /* Write Data to DATA Register for conversion(DATA[9:0]) */
-    DAC_REGS->DAC_DATA = DAC_DATA_MSB_MASK & DAC_DATA_DATA(data);
+    DAC_REGS->DAC_DATA = (uint16_t)(DAC_DATA_MSB_MASK & DAC_DATA_DATA((uint32_t)data));
     while((DAC_REGS->DAC_SYNCBUSY & DAC_SYNCBUSY_DATA_Msk) == DAC_SYNCBUSY_DATA_Msk)
     {
         /* Wait for Synchronization after writing Data to DATA Register */
